+++ tests/promises/nested.validate.js
@@ -4,7 +4,7 @@ export default ($) => {
   describe('Form validate()', () => {
     // $R
     it('$R validate() should be true', (done) => {
-      $.$R.validate().then((isValid) => {
+      $.$R.validate().then(({ isValid }) => {
         expect(isValid).to.be.true; // eslint-disable-line
         done();
       });
@@ -12,7 +12,7 @@ export default ($) => {
 
     // $S
     it('$S validate() should be false', (done) => {
-      $.$S.validate().then((isValid) => {
+      $.$S.validate().then(({ isValid }) => {
         expect(isValid).to.be.false; // eslint-disable-line
         done();
       });